requirements.txt
flask
ultralytics
opencv-python
face-recognition
mediapipe
scipy
mysql-connector-python
numpy
deep-sort-realtime